tests/wpt/web-platform-tests/offscreen-canvas/pixel-manipulation/2d.imageData.object.properties.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.imageData.object.properties</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>2d.imageData.object.properties</h1>
<p class="desc">ImageData objects have the right properties</p>


<script>
var t = async_test("ImageData objects have the right properties");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var imgdata = ctx.getImageData(0, 0, 10, 10);
_assertSame(typeof(imgdata.width), 'number', "typeof(imgdata.width)", "'number'");
_assertSame(typeof(imgdata.height), 'number', "typeof(imgdata.height)", "'number'");
_assertSame(typeof(imgdata.data), 'object', "typeof(imgdata.data)", "'object'");
t.done();

});
</script>
